ecommerce_backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
gunicorn>=20.1.0 